*Rename pil/dashboard to pil/application *Replace license with licence *Replace copy on "Do you have any module exemptions?" page
