+++ pages/common/content/index.js
@@ -26,7 +26,7 @@ module.exports = {
       confirm: 'Confirm changes'
     },
     pil: {
-      categories: 'Choose personal license (PIL) category',
+      categories: 'Choose personal licence (PIL) category',
       dashboard: 'Apply for categories A, B, C, D and F PIL',
       categoryAToF: 'Categories A,B,C,D and F',
       categoryE: 'Category E'
